Move OS-specific funcs to path.go
